+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/VerikDomainObject.kt
@@ -14,7 +14,9 @@ interface VerikDomainObject {
 
     fun import(configure: VerikImportDomainObject.() -> Unit)
     fun compile(configure: VerikCompileDomainObject.() -> Unit)
+
     fun dsim(configure: DsimTargetDomainObject.() -> Unit)
     fun iverilog(configure: IverilogTargetDomainObject.() -> Unit)
     fun vivado(configure: VivadoTargetDomainObject.() -> Unit)
+    fun xrun(configure: XrunTargetDomainObject.() -> Unit)
 }