+++ verik-plugin/src/main/kotlin/io/verik/plugin/target/XrunTargetBuilder.kt
@@ -5,34 +5,31 @@
 package io.verik.plugin.target
 
 import io.verik.compiler.common.TextFile
-import io.verik.plugin.config.IverilogTargetConfig
+import io.verik.plugin.config.XrunTargetConfig
 
 /**
- * Builder for Icarus Verilog targets
+ * Builder for Cadence xrun targets.
  */
-object IverilogTargetBuilder {
+object XrunTargetBuilder {
 
-    fun build(targetConfig: IverilogTargetConfig): TextFile {
+    fun build(targetConfig: XrunTargetConfig): TextFile {
         val path = targetConfig.buildDir.resolve("Makefile")
         val header = TargetFileHeaderBuilder.build(targetConfig, path)
 
         val builder = StringBuilder()
         builder.appendLine(header)
-        builder.appendLine("IVERILOG = iverilog")
-        builder.appendLine("VVP = vvp")
+        builder.appendLine("XRUN = xrun")
         builder.appendLine()
 
         builder.appendLine(".PHONY: clean all")
         builder.appendLine()
 
         builder.appendLine("all:")
-        builder.append("\t\$(IVERILOG) -o out -g2012 -s ${targetConfig.top}")
-        builder.appendLine(" -I ../compile ../compile/out.sv")
-        builder.appendLine("\t\$(VVP) -l log out")
+        builder.appendLine("\t\$(XRUN) -top ${targetConfig.top} -incdir ../compile ../compile/out.sv")
         builder.appendLine()
 
         builder.appendLine("clean:")
-        builder.appendLine("\trm -rf out log")
+        builder.appendLine("\trm -rf xcelium.d xrun.history xrun.log")
 
         return TextFile(path, builder.toString())
     }