+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add the Metrics dsim generation target.
 - Add the Xilinx vivado generation target.
 - Add the Icarus Verilog generation target.
+- Add the Cadence xrun generation target.
 
 ## [0.1.15]
 ### Compiler